Fixes all previous uses of BigInt in Earnings, which was throwing an error when FMP tried to update

+++ server/models/earnings.js
@@ -13,13 +13,13 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     revenue: {
-      type: DataTypes.BIGINT,
+      type: DataTypes.NUMERIC,
     },
     costOfRevenue: {
       type: DataTypes.NUMERIC,
     },
     grossProfit: {
-      type: DataTypes.BIGINT,
+      type: DataTypes.NUMERIC,
     },
     grossProfitRatio: {
       type: DataTypes.NUMERIC,
